fix(resource): 修复删除页面集时的空指针异常
- 增加对应用菜单对象的空值检查
- 在删除页面集前增加对页面集对象的空值检查
- 优化删除逻辑，避免在页面ID列表为空时执行删除操作
- 调整代码执行顺序，确保先验证再执行删除操作

+++ onebase-module-app/onebase-module-app-build/src/main/java/com/cmsr/onebase/module/app/build/service/resource/PageSetServiceImpl.java
@@ -138,15 +138,22 @@ public class PageSetServiceImpl implements PageSetService {
     @Transactional(rollbackFor = Exception.class)
     public void deletePageSetByMenuId(Long menuId) {
         AppMenuDO appMenuDO = appMenuRepository.getById(menuId);
-        // 找到页面集
-        AppResourcePagesetDO pageSetDO = pageSetRepository.findPageSetByAppIdAndMenuUuid(appMenuDO.getApplicationId(), appMenuDO.getMenuUuid());
+        if (appMenuDO == null) {
+            return;
+        }
+        appMenuRepository.removeById(appMenuDO);
         // 删除页面集关联的页面
-        List<Long> pageIds = pageRepository.findIdsByAppIdAndPageSetUuid(appMenuDO.getApplicationId(), pageSetDO.getPageSetUuid());
-        // 删除页面集
+        AppResourcePagesetDO pageSetDO = pageSetRepository.findPageSetByAppIdAndMenuUuid(appMenuDO.getApplicationId(), appMenuDO.getMenuUuid());
+        if (pageSetDO == null) {
+            return;
+        }
         pageSetRepository.removeById(pageSetDO);
         // 删除页面
+        List<Long> pageIds = pageRepository.findIdsByAppIdAndPageSetUuid(appMenuDO.getApplicationId(), pageSetDO.getPageSetUuid());
+        if (CollectionUtils.isEmpty(pageIds)) {
+            return;
+        }
         pageRepository.removeByIds(pageIds);
-
     }
 
     @Override
